+++ tdengine30reader/src/test/java/com/alibaba/datax/plugin/tdengine30reader/TDengine2StreamTest.java
@@ -1,4 +1,4 @@
-package com.alibaba.datax.plugin.tdengine20reader;
+package com.alibaba.datax.plugin.tdengine30reader;
 
 import com.alibaba.datax.core.Engine;
 import org.junit.Ignore;
@@ -11,47 +11,36 @@ import java.sql.Statement;
 import java.util.Random;
 
 @Ignore
-public class TDengine2DMTest {
-    private static final String host1 = "192.168.56.105";
-    private static final String host2 = "192.168.0.72";
+public class TDengine2StreamTest {
 
-    private final Random random = new Random(System.currentTimeMillis());
+    private static final String host = "192.168.56.105";
+    private static final Random random = new Random(System.currentTimeMillis());
 
     @Test
-    public void t2dm_case01() throws Throwable {
+    public void case01() throws Throwable {
         // given
-        createSupTable("ms");
+        prepare("ms");
 
         // when
-        String[] params = {"-mode", "standalone", "-jobid", "-1", "-job", "src/test/resources/t2dm.json"};
+        String[] params = {"-mode", "standalone", "-jobid", "-1", "-job", "src/test/resources/t2stream-1.json"};
         System.setProperty("datax.home", "../target/datax/datax");
         Engine.entry(params);
     }
 
     @Test
-    public void t2dm_case02() throws Throwable {
+    public void case02() throws Throwable {
         // given
-        createSupTable("us");
+        prepare("ms");
 
         // when
-        String[] params = {"-mode", "standalone", "-jobid", "-1", "-job", "src/test/resources/t2dm.json"};
+        String[] params = {"-mode", "standalone", "-jobid", "-1", "-job", "src/test/resources/t2stream-2.json"};
         System.setProperty("datax.home", "../target/datax/datax");
         Engine.entry(params);
     }
 
-    @Test
-    public void t2dm_case03() throws Throwable {
-        // given
-        createSupTable("ns");
 
-        // when
-        String[] params = {"-mode", "standalone", "-jobid", "-1", "-job", "src/test/resources/t2dm.json"};
-        System.setProperty("datax.home", "../target/datax/datax");
-        Engine.entry(params);
-    }
-
-    private void createSupTable(String precision) throws SQLException {
-        final String url = "jdbc:TAOS-RS://" + host1 + ":6041/";
+    private void prepare(String precision) throws SQLException {
+        final String url = "jdbc:TAOS-RS://" + host + ":6041/";
         try (Connection conn = DriverManager.getConnection(url, "root", "taosdata")) {
             Statement stmt = conn.createStatement();
 
@@ -71,16 +60,7 @@ public class TDengine2DMTest {
             }
             stmt.close();
         }
-
-        final String url2 = "jdbc:dm://" + host2 + ":5236";
-        try (Connection conn = DriverManager.getConnection(url2, "TESTUSER", "test123456")) {
-            conn.setAutoCommit(true);
-            Statement stmt = conn.createStatement();
-            stmt.execute("drop table if exists stb2");
-            stmt.execute("create table stb2(ts timestamp, f1 tinyint, f2 smallint, f3 int, f4 bigint, f5 float, " +
-                    "f6 double, f7 BIT, f8 VARCHAR(100), f9 VARCHAR2(200), t1 timestamp, t2 tinyint, t3 smallint, " +
-                    "t4 int, t5 bigint, t6 float, t7 double, t8 BIT, t9 VARCHAR(100), t10 VARCHAR2(200))");
-        }
     }
+
 
 }
