--- tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine20writer/DataHandler.java
@@ -1,8 +1,0 @@
-package com.alibaba.datax.plugin.writer.tdengine20writer;
-
-import com.alibaba.datax.common.plugin.RecordReceiver;
-import com.alibaba.datax.common.plugin.TaskPluginCollector;
-
-public interface DataHandler {
-    int handle(RecordReceiver lineReceiver, TaskPluginCollector collector);
-}
