--- tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine20writer/TableMeta.java
@@ -1,22 +1,0 @@
-package com.alibaba.datax.plugin.writer.tdengine20writer;
-
-public class TableMeta {
-    TableType tableType;
-    String tbname;
-    int columns;
-    int tags;
-    int tables;
-    String stable_name;
-
-    @Override
-    public String toString() {
-        return "TableMeta{" +
-                "tableType=" + tableType +
-                ", tbname='" + tbname + '\'' +
-                ", columns=" + columns +
-                ", tags=" + tags +
-                ", tables=" + tables +
-                ", stable_name='" + stable_name + '\'' +
-                '}';
-    }
-}
