--- tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine20writer/TableType.java
@@ -1,5 +1,0 @@
-package com.alibaba.datax.plugin.writer.tdengine20writer;
-
-public enum TableType {
-    SUP_TABLE, SUB_TABLE, NML_TABLE
-}
