--- tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine20writer/TimestampPrecision.java
@@ -1,5 +1,0 @@
-package com.alibaba.datax.plugin.writer.tdengine20writer;
-
-public enum TimestampPrecision {
-    MILLISEC, MICROSEC, NANOSEC
-}
