+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/ColumnMeta.java
@@ -1,0 +1,24 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+public class ColumnMeta {
+    String field;
+    String type;
+    int length;
+    String note;
+    boolean isTag;
+    boolean isPrimaryKey;
+    Object value;
+
+    @Override
+    public String toString() {
+        return "ColumnMeta{" +
+                "field='" + field + '\'' +
+                ", type='" + type + '\'' +
+                ", length=" + length +
+                ", note='" + note + '\'' +
+                ", isTag=" + isTag +
+                ", isPrimaryKey=" + isPrimaryKey +
+                ", value=" + value +
+                '}';
+    }
+}
