+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/DataHandler.java
@@ -1,0 +1,8 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+import com.alibaba.datax.common.plugin.RecordReceiver;
+import com.alibaba.datax.common.plugin.TaskPluginCollector;
+
+public interface DataHandler {
+    int handle(RecordReceiver lineReceiver, TaskPluginCollector collector);
+}
