+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/Key.java
@@ -1,0 +1,18 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+public class Key {
+    public static final String USERNAME = "username";
+    public static final String PASSWORD = "password";
+    public static final String CONNECTION = "connection";
+    public static final String BATCH_SIZE = "batchSize";
+    public static final String TABLE = "table";
+    public static final String JDBC_URL = "jdbcUrl";
+    public static final String COLUMN = "column";
+    public static final String IGNORE_TAGS_UNMATCHED = "ignoreTagsUnmatched";
+
+    public static final String BEGIN_DATETIME = "beginDateTime";
+    public static final String END_DATETIME = "endDateTime";
+    public static final String WHERE = "where";
+    public static final String QUERY_SQL = "querySql";
+    public static final String MANDATORY_ENCODING = "mandatoryEncoding";
+}
