+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/TableMeta.java
@@ -1,0 +1,22 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+public class TableMeta {
+    TableType tableType;
+    String tbname;
+    int columns;
+    int tags;
+    int tables;
+    String stable_name;
+
+    @Override
+    public String toString() {
+        return "TableMeta{" +
+                "tableType=" + tableType +
+                ", tbname='" + tbname + '\'' +
+                ", columns=" + columns +
+                ", tags=" + tags +
+                ", tables=" + tables +
+                ", stable_name='" + stable_name + '\'' +
+                '}';
+    }
+}
