+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/TableType.java
@@ -1,0 +1,5 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+public enum TableType {
+    SUP_TABLE, SUB_TABLE, NML_TABLE
+}
