+++ tdengine30writer/src/main/java/com/alibaba/datax/plugin/writer/tdengine30writer/TimestampPrecision.java
@@ -1,0 +1,5 @@
+package com.alibaba.datax.plugin.writer.tdengine30writer;
+
+public enum TimestampPrecision {
+    MILLISEC, MICROSEC, NANOSEC
+}
